Extract MOSAIC usage data level 0 from Evergreen
Conifer's two participants in the SPLURGE hackfest are the University of
Windsor and Laurentian University.

The randomization might not be all that random, come to think of it;
computers could determine what timestamp value would generate all
integers for the desired output. Will go back to the drawing board on
that one.

Signed-off-by: Dan Scott <dan@coffeecode.net>